requirements.txt
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
django-environ==0.4.5